requirements.txt
# Core Django
django
djangorestframework
mysqlclient

# Security & Tokens
djangorestframework-simplejwt

# Utilities & Config
django-environ
gunicorn

# Documentation
drf-spectacular

# Testing & Quality
pytest-django
flake8

# External API
requests